Moves common command infrastructure to new namespace
